# frontend/src/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Afisharr",
    "description": "Plex collections, posters, and overlay manager.",
    "contact": {
      "name": "Afisharr contributors"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Signs in with a username and a password.",
        "description": "Both limits are taken before the hash and handed back after it, and only on\nsuccess: a limit that kept counting after a correct password would lock out\nthe operator signing in from a fourth device (PRD §21.4.3).\n\nThat ordering is the whole of whether the limit is a bound. Read the other\nway round — consult the counter, hash, count the failure afterwards — an\nattempt that has not finished has not failed, so a burst arriving inside one\ninstant reads the same empty counter and every guess in it runs. The\nsemaphore in `accounts::verify` bounds the *memory* that costs, four\nArgon2id operations at a time, and bounds nothing about how many guesses the\naccount gives up: the rest queue and run in turn. Taking the attempt first\nis what makes five guesses five guesses.",
        "operationId": "log_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedIn"
                }
              }
            }
          },
          "400": {
            "description": "The request body was not readable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "The credentials were not accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Setup has not been completed on this instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Signs out, revoking the session that made the request.",
        "operationId": "log_out",
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "No session to sign out of",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Setup has not been completed on this instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/plex/pin": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Starts a Plex sign-in.",
        "operationId": "start_plex_pin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A pin was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinStarted"
                }
              }
            }
          },
          "400": {
            "description": "The request body was not readable, or the return target is not this instance's configured origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Setup has not been completed on this instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "plex.tv issued the sign-in under a different client identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "plex.tv could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/plex/pin/{id}": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Polls a Plex sign-in, and signs the operator in when the token arrives.",
        "operationId": "poll_plex_pin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The attempt returned by the start call",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The attempt's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinState"
                }
              }
            }
          },
          "403": {
            "description": "The attempt belongs to another browser, the Plex account is not linked to an Afisharr account, or setup has not been completed on this instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The client identifier changed, or the attempt was already completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many calls to plex.tv",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "plex.tv could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/session": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Reports the signed-in account.",
        "operationId": "whoami",
        "responses": {
          "200": {
            "description": "The signed-in account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedIn"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Setup has not been completed on this instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Lists one directory inside one root.",
        "description": "Every refusal names the root and never the resolved path (`I-SEC-3`).",
        "operationId": "browse",
        "parameters": [
          {
            "name": "root",
            "in": "path",
            "description": "The identifier of the root to walk.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "path",
            "description": "The path inside the root. Empty means the root itself.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The directory's contents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryListing"
                }
              }
            }
          },
          "400": {
            "description": "The query was not readable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No accepted credential was presented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The path is not inside the root, that account does not administer this instance, that key was not issued with the scope this route needs, or setup has not been completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such root",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/roots": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Lists the roots the operator has enabled.",
        "operationId": "roots",
        "responses": {
          "200": {
            "description": "Every enabled root",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RootView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No accepted credential was presented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "That account does not administer this instance, that key was not issued with the scope this route needs, or setup has not been completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/api-keys/{id}": {
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Revokes one key. It is refused on its next use.",
        "operationId": "revoke",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The key to revoke",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The key is revoked"
          },
          "401": {
            "description": "No accepted credential was presented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "That account does not administer this instance, that key was not issued with the scope this route needs, or setup has not been completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such key, or it was already revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Credentials": {
        "type": "object",
        "description": "What the sign-in form sends.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "The password."
          },
          "username": {
            "type": "string",
            "description": "The account name."
          }
        },
        "additionalProperties": false
      },
...
      "SignedIn": {
        "type": "object",
        "description": "Who the caller now is.",
        "required": [
          "userId",
          "username",
          "isAdmin"
        ],
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ],
            "description": "The name to show, when it differs from the account name."
          },
          "isAdmin": {
            "type": "boolean",
            "description": "Whether the account holds administrator rights."
          },
          "userId": {
            "type": "string",
            "description": "The account's identifier."
          },
          "username": {
            "type": "string",
            "description": "The account name."
          }
        }
      },
      "StartPin": {
        "type": "object",
        "description": "What the interface asks for when it starts a Plex sign-in.",
        "required": [
          "oauth"
        ],
        "properties": {
          "forwardUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where plex.tv returns the operator, for the OAuth variant.\n\nIt must be on this instance's configured `publicOrigin`. plex.tv\nredirects to whatever this asks for, so a target anywhere else turns the\nendpoint into a redirector wearing a legitimate `app.plex.tv/auth` URL,\nand the operator who completes the sign-in lands on somebody else's\npage."
          },
          "oauth": {
            "type": "boolean",
            "description": "`pin` shows a four-character code; `oauth` sends the operator to\nplex.tv's hosted sign-in."
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}